requirements.txt
# Core LLM
google-generativeai>=0.5.0

# Parsing
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
paddleocr>=2.7.0         # optional, for OCR diagrams
layoutparser>=0.3.4      # optional, structured parsing

# Utilities
pyyaml>=6.0
jinja2>=3.1.2
pydantic>=2.7.0

# Testing & Dev
pytest>=8.1.0
black>=24.4.0
isort>=5.13.0
